+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## Added
+- Can now press `t`/`T` to create a new tab (or restart current tab) using the current/selected object(s) as the seed object.
+to give a console instance variable name to current objects without leaving scry session.
 
 ## Changed